XC-317: Clippy
