+++ canister/scripts/examples.sh
@@ -61,8 +61,8 @@ GET_BLOCK_PARAMS="(
 )"
 CYCLES=$(dfx canister call sol_rpc getBlockCyclesCost "$GET_BLOCK_PARAMS" $FLAGS --output json | jq '.Ok' --raw-output || exit 1)
 GET_BLOCK_OUTPUT=$(dfx canister call sol_rpc getBlock "$GET_BLOCK_PARAMS" $FLAGS --output json --with-cycles "$CYCLES" || exit 1)
-FIRST_SIGNATURE=$(jq --raw-output '.Consistent.Ok[0].signatures[0][0]' <<< "$GET_BLOCK_OUTPUT")
-SECOND_SIGNATURE=$(jq --raw-output '.Consistent.Ok[0].signatures[0][1]' <<< "$GET_BLOCK_OUTPUT")
+FIRST_SIGNATURE=$(jq --raw-output '.Consistent.Ok[0].signature[0][0]' <<< "$GET_BLOCK_OUTPUT")
+SECOND_SIGNATURE=$(jq --raw-output '.Consistent.Ok[0].signature[0][1]' <<< "$GET_BLOCK_OUTPUT")
 
 # Fetch the statuses of the first two transactions in the received block
 GET_SIGNATURE_STATUSES_PARAMS="(
