+++ examples/basic_solana/src/solana_wallet.rs
@@ -9,8 +9,8 @@ use crate::{
     state::{lazy_call_ed25519_public_key, read_state},
 };
 use candid::Principal;
-use sol_rpc_client::{threshold_signatures::sign_transaction, IcRuntime};
-use sol_rpc_types::{DerivationPath, SignTransactionRequestParams};
+use sol_rpc_client::{threshold_sig::sign_transaction, IcRuntime};
+use sol_rpc_types::solana::::DerivationPath;
 use solana_message::Message;
 use solana_pubkey::Pubkey;
 use solana_signature::Signature;
