+++ libs/types/src/lib.rs
@@ -25,13 +25,13 @@ use serde::Serialize;
 pub use solana::{
     account::{AccountData, AccountEncoding, AccountInfo, ParsedAccount},
     request::{
-        CommitmentLevel, DataSlice, DerivationPath, Ed25519KeyId, GetAccountInfoEncoding,
-        GetAccountInfoParams, GetBalanceParams, GetBlockCommitmentLevel, GetBlockParams,
-        GetRecentPrioritizationFeesParams, GetSignatureStatusesParams, GetSlotParams,
-        GetTokenAccountBalanceParams, GetTransactionEncoding, GetTransactionParams,
-        SendTransactionEncoding, SendTransactionParams, SignTransactionRequestParams,
-        TransactionDetails,
+        CommitmentLevel, DataSlice, GetAccountInfoEncoding, GetAccountInfoParams, GetBalanceParams,
+        GetBlockCommitmentLevel, GetBlockParams, GetRecentPrioritizationFeesParams,
+        GetSignatureStatusesParams, GetSlotParams, GetTokenAccountBalanceParams,
+        GetTransactionEncoding, GetTransactionParams, SendTransactionEncoding,
+        SendTransactionParams, TransactionDetails,
     },
+    signature::{DerivationPath, Ed25519KeyId},
     transaction::{
         error::{InstructionError, TransactionError},
         instruction::{CompiledInstruction, InnerInstructions, Instruction},
